CSS-color basic
Color coding is an important element in web design. Colors determine the feel and overall look of a web page. Coding your colors with CSS allows you to control the style and design of your web page. Also, by color coding with CSS, you can ensure that colors are used harmoniously and consistently on a web page.